requirements.txt
colorama==0.4.6
Faker==18.10.1
folium==0.14.0
geopy==2.3.0
holidays==0.26
kafka_python==2.0.2
keras==2.12.0
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.2
pymongo==4.3.3
pyspark==3.4.0
pyTelegramBotAPI==4.12.0
pyTelegramBotAPI==4.12.0
redis==4.5.5
Requests==2.31.0
scikit_learn==1.2.2
torch==2.0.1
